Fix modifications of settings & saving
Since the settings are accumulated in a ChainMap, they would have to
either be accessed via `settings.maps[1].save()` or via an explicitly
named global.

I chose the latter, since it does not rely on the `ChainMap` order and
is more explicit.

The same also applies to writing to the `ChainMap` before, of course.